Verify provided index is a valid Json Token
Previously we did not verify the index provided to any of the Json
functions, which could lead to the program blowing up if a bad one was
provided (since it would be passed to the _Mem functions typically, and
give an out-of-bounds error). This fixes that so that a JsonHadError and
JsonError is provided about the invalid token index.