Don't consider zero a valid token
Because variables default to zero, zero is commonly passed to
qb64pe-json API functions by mistake when there are errors in the code.
Currently we will use the zero token as a valid token, which means these
don't result in errors but instead just mess-up the Json structure in
hard to detect ways, which isn't good.

Skipping the zero token instead is much more sensible, since as the cost
of a small amount of memory we can avoid one of the most common sources
of errors.